This must be what Jesse is talking about, it is a large UTF-8 decode
This must be what Jesse is talking about, it is a large UTF-8 decode
where I copied the technique from OpenJDKs technique. This was not used
in any of the benchmarks that were published as we did not publish the
results of the byte[] parser because it only beats Jackson and GSON in
rare circumstances and never does as well as the char[] based parsers.
Also the UTF-8 decode seems much slower than that which comes with the
JDK so their is no point in having this code. 1) Jesse’s complaint 2)
not used by any of the fast parsers 3) it is slower than what the JDK
provides