Jesse complains of copied JDK code, which turns out to be copied JavaDoc.
Jesse complains of copied JDK code, which turns out to be copied
JavaDoc.
I am removing the JavaDoc. I should not have copied the JavaDoc.
I remember doing this and thinking what a clever way to write JavaDocs
of a public lib.
There is no code copied that I know of. I subclassed PrintWriter.
CharBuf is a StringBuilder like class that also looks like a
PrintWriter.
These methods make up a very small fraction of what CharBuf does and
are merely here
to allow CharBuf to be used as PrintWriter.